Refactor code structure
Improved the organization and readability of the codebase by refactoring various components and utilities. This change aims to enhance maintainability and facilitate future development.
[skip gpt_engineer]
